pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "cybereye4-wrapper"
version = "0.2.0"
description = "Knowledge-augmented Q&A and decision wrapper for Cybereye4"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
    "openai>=1.0",
    "neo4j>=5.0,<6",
    "python-dotenv>=1.0.0",
    "opencv-python-headless>=4.8.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
